requirements.txt
pandas==1.5.3
numpy==1.23.5
scipy==1.10.1
scikit-learn==1.2.2
matplotlib==3.7.3
tensorflow==2.10.1
aiohttp==3.9.5
statsmodels==0.14.2
PyYAML==6.0.1
seaborn==0.12.2
fastcluster==1.2.6
yfinance==0.2.48
openai>=1.0.0,<2.0.0
python-telegram-bot>=13.0,<21.0
